Refactor bottom panel
